Stub out noosphere-name-system on wasm32-unknown-unknown builds.
