+++ rust/Cargo.toml
@@ -11,6 +11,7 @@ members = [
   "noosphere-cli",
   "noosphere-name-system",
 ]
+
 # See: https://github.com/rust-lang/rust/issues/90148#issuecomment-949194352
 resolver = "2"
